Update of all dockerfiles (see full msg)
All dockerfiles updated with new dev_env. Neovim is now only used to do
a headless vim install, but vim is the main IDE. This was chosen due to
numerous issues encountered with neovim during development of projects.
YCM is compiled but without go support due to issues building a go
dependency. The rust installer was updated to be inline with the new
rustup url.